sm8250-common: props: Disable Vsync for CPU Rendered-apps.
* Choreographer VSync has some performance impact for CPU rendered apps
as the app has to wait for VSync to start a new frame draw. When the
frame draw time exceeds one VSync, this has a direct impact on the FPS.
debug.cpurend.vsync property controls the VSync for CPU rendered apps.

Disable VSync for CPU rendered app, by setting this property to false.

Signed-off-by: namaskaragusthi <senaoutsider123@gmail.com>